Implement the trap manager and move watermarks etc out of the clocking code
