+++ arctic/__init__.py
@@ -1,3 +1,9 @@
 from arctic.clock import Clocker
-from arctic.model import ArcticParams, CCDVolume, CCDVolumeComplex, Trap
+from arctic.model import (
+    ArcticParams,
+    CCDVolume,
+    CCDVolumeComplex,
+    Trap,
+    TrapManager,
+)
 from arctic import util
