+++ test_arctic/unit/test_clock.py
@@ -5,593 +5,6 @@ import arctic as ac
 from arctic import clock
 
 
-class TestElectronsReleasedAndUpdatedWatermarks:
-    def test__empty_release(self):
-
-        traps = [ac.Trap(density=10, lifetime=-1 / np.log(0.5))]
-
-        clocker = ac.Clocker()
-        watermarks = clocker.initial_watermarks_from_rows_and_total_traps(
-            rows=6, total_traps=1
-        )
-
-        (
-            electrons_released,
-            watermarks,
-        ) = clock.electrons_released_in_pixel_and_updated_watermarks(
-            watermarks=watermarks, traps=traps
-        )
-
-        assert electrons_released == pytest.approx(0)
-        assert watermarks == pytest.approx(
-            np.array([[0, 0], [0, 0], [0, 0], [0, 0], [0, 0], [0, 0]])
-        )
-
-    def test__single_trap(self):
-
-        traps = [ac.Trap(density=10, lifetime=-1 / np.log(0.5))]
-
-        watermarks = np.array(
-            [[0.5, 0.8], [0.2, 0.4], [0.1, 0.2], [0, 0], [0, 0], [0, 0]]
-        )
-
-        (
-            electrons_released,
-            watermarks,
-        ) = clock.electrons_released_in_pixel_and_updated_watermarks(
-            watermarks=watermarks, traps=traps
-        )
-
-        assert electrons_released == pytest.approx(2.5)
-        assert watermarks == pytest.approx(
-            np.array(
-                [[0.5, 0.4], [0.2, 0.2], [0.1, 0.1], [0, 0], [0, 0], [0, 0]]
-            )
-        )
-
-    def test__multiple_traps(self):
-
-        traps = [
-            ac.Trap(density=10, lifetime=-1 / np.log(0.5)),
-            ac.Trap(density=8, lifetime=-1 / np.log(0.2)),
-        ]
-
-        watermarks = np.array(
-            [
-                [0.5, 0.8, 0.3],
-                [0.2, 0.4, 0.2],
-                [0.1, 0.2, 0.1],
-                [0, 0, 0],
-                [0, 0, 0],
-                [0, 0, 0],
-            ]
-        )
-
-        (
-            electrons_released,
-            watermarks,
-        ) = clock.electrons_released_in_pixel_and_updated_watermarks(
-            watermarks=watermarks, traps=traps
-        )
-
-        assert electrons_released == pytest.approx(2.5 + 1.28)
-        assert watermarks == pytest.approx(
-            np.array(
-                [
-                    [0.5, 0.4, 0.06],
-                    [0.2, 0.2, 0.04],
-                    [0.1, 0.1, 0.02],
-                    [0, 0, 0],
-                    [0, 0, 0],
-                    [0, 0, 0],
-                ]
-            )
-        )
-
-
-class TestElectronsCapturedByTraps:
-    def test__first_capture(self):
-
-        clocker = ac.Clocker()
-
-        traps = [ac.Trap(density=10, lifetime=-1 / np.log(0.5))]
-
-        watermarks = clocker.initial_watermarks_from_rows_and_total_traps(
-            rows=6, total_traps=1
-        )
-        electron_fractional_height = 0.5
-
-        electrons_captured = clock.electrons_captured_by_traps(
-            electron_fractional_height=electron_fractional_height,
-            watermarks=watermarks,
-            traps=traps,
-        )
-
-        assert electrons_captured == pytest.approx(0.5 * 10)
-
-    def test__new_highest_watermark(self):
-
-        traps = [ac.Trap(density=10, lifetime=-1 / np.log(0.5))]
-
-        watermarks = np.array(
-            [[0.5, 0.8], [0.2, 0.4], [0.1, 0.3], [0, 0], [0, 0], [0, 0]]
-        )
-        electron_fractional_height = 0.9
-
-        electrons_captured = clock.electrons_captured_by_traps(
-            electron_fractional_height=electron_fractional_height,
-            watermarks=watermarks,
-            traps=traps,
-        )
-
-        assert electrons_captured == pytest.approx(
-            (0.1 + 0.12 + 0.07 + 0.1) * 10
-        )
-
-    def test__middle_new_watermarks(self):
-
-        traps = [ac.Trap(density=10, lifetime=-1 / np.log(0.5))]
-
-        watermarks = np.array(
-            [[0.5, 0.8], [0.2, 0.4], [0.1, 0.3], [0, 0], [0, 0], [0, 0]]
-        )
-        electron_fractional_height = 0.6
-
-        electrons_captured = clock.electrons_captured_by_traps(
-            electron_fractional_height=electron_fractional_height,
-            watermarks=watermarks,
-            traps=traps,
-        )
-
-        assert electrons_captured == pytest.approx((0.1 + 0.06) * 10)
-
-        watermarks = np.array(
-            [[0.5, 0.8], [0.2, 0.4], [0.1, 0.3], [0, 0], [0, 0], [0, 0]]
-        )
-        electron_fractional_height = 0.75
-
-        electrons_captured = clock.electrons_captured_by_traps(
-            electron_fractional_height=electron_fractional_height,
-            watermarks=watermarks,
-            traps=traps,
-        )
-
-        assert electrons_captured == pytest.approx((0.1 + 0.12 + 0.035) * 10)
-
-    def test__new_lowest_watermark(self):
-
-        traps = [ac.Trap(density=10, lifetime=-1 / np.log(0.5))]
-
-        watermarks = np.array(
-            [[0.5, 0.8], [0.2, 0.4], [0.1, 0.3], [0, 0], [0, 0], [0, 0]]
-        )
-        electron_fractional_height = 0.3
-
-        electrons_captured = clock.electrons_captured_by_traps(
-            electron_fractional_height=electron_fractional_height,
-            watermarks=watermarks,
-            traps=traps,
-        )
-
-        assert electrons_captured == pytest.approx(0.3 * 0.2 * 10)
-
-    def test__multiple_traps(self):
-
-        traps = [
-            ac.Trap(density=10, lifetime=-1 / np.log(0.5)),
-            ac.Trap(density=8, lifetime=-1 / np.log(0.2)),
-        ]
-
-        watermarks = np.array(
-            [
-                [0.5, 0.8, 0.7],
-                [0.2, 0.4, 0.3],
-                [0.1, 0.3, 0.2],
-                [0, 0, 0],
-                [0, 0, 0],
-                [0, 0, 0],
-            ]
-        )
-
-        electron_fractional_height = 0.6
-
-        electrons_captured = clock.electrons_captured_by_traps(
-            electron_fractional_height=electron_fractional_height,
-            watermarks=watermarks,
-            traps=traps,
-        )
-
-        assert electrons_captured == pytest.approx(
-            (0.1 + 0.06) * 10 + (0.15 + 0.07) * 8
-        )
-
-
-class TestUpdateWatermarks:
-    def test__first_capture(self):
-
-        clocker = ac.Clocker()
-
-        watermarks = clocker.initial_watermarks_from_rows_and_total_traps(
-            rows=6, total_traps=1
-        )
-        electron_fractional_height = 0.5
-
-        watermarks = clock.updated_watermarks_from_capture(
-            electron_fractional_height=electron_fractional_height,
-            watermarks=watermarks,
-        )
-
-        assert watermarks == pytest.approx(
-            np.array([[0.5, 1], [0, 0], [0, 0], [0, 0], [0, 0], [0, 0]])
-        )
-
-    def test__new_highest_watermark(self):
-
-        watermarks = np.array(
-            [[0.5, 0.8], [0.2, 0.4], [0.1, 0.3], [0, 0], [0, 0], [0, 0]]
-        )
-        electron_fractional_height = 0.9
-
-        watermarks = clock.updated_watermarks_from_capture(
-            electron_fractional_height=electron_fractional_height,
-            watermarks=watermarks,
-        )
-
-        assert watermarks == pytest.approx(
-            np.array([[0.9, 1], [0, 0], [0, 0], [0, 0], [0, 0], [0, 0]])
-        )
-
-    def test__middle_new_watermark(self):
-
-        watermarks = np.array(
-            [[0.5, 0.8], [0.2, 0.4], [0.1, 0.3], [0, 0], [0, 0], [0, 0]]
-        )
-        electron_fractional_height = 0.6
-
-        watermarks = clock.updated_watermarks_from_capture(
-            electron_fractional_height=electron_fractional_height,
-            watermarks=watermarks,
-        )
-
-        assert watermarks == pytest.approx(
-            np.array([[0.6, 1], [0.1, 0.4], [0.1, 0.3], [0, 0], [0, 0], [0, 0]])
-        )
-
-        watermarks = np.array(
-            [[0.5, 0.8], [0.2, 0.4], [0.1, 0.3], [0, 0], [0, 0], [0, 0]]
-        )
-        electron_fractional_height = 0.75
-
-        watermarks = clock.updated_watermarks_from_capture(
-            electron_fractional_height=electron_fractional_height,
-            watermarks=watermarks,
-        )
-
-        assert watermarks == pytest.approx(
-            np.array([[0.75, 1], [0.05, 0.3], [0, 0], [0, 0], [0, 0], [0, 0]])
-        )
-
-    def test__new_lowest_watermark(self):
-
-        watermarks = np.array(
-            [[0.5, 0.8], [0.2, 0.4], [0.1, 0.3], [0, 0], [0, 0], [0, 0]]
-        )
-        electron_fractional_height = 0.3
-
-        watermarks = clock.updated_watermarks_from_capture(
-            electron_fractional_height=electron_fractional_height,
-            watermarks=watermarks,
-        )
-
-        assert watermarks == pytest.approx(
-            np.array(
-                [[0.3, 1], [0.2, 0.8], [0.2, 0.4], [0.1, 0.3], [0, 0], [0, 0]]
-            )
-        )
-
-    def test__multiple_traps(self):
-
-        watermarks = np.array(
-            [
-                [0.5, 0.8, 0.7, 0.6],
-                [0.2, 0.4, 0.3, 0.2],
-                [0.1, 0.3, 0.2, 0.1],
-                [0, 0, 0, 0],
-                [0, 0, 0, 0],
-                [0, 0, 0, 0],
-            ]
-        )
-        electron_fractional_height = 0.6
-
-        watermarks = clock.updated_watermarks_from_capture(
-            electron_fractional_height=electron_fractional_height,
-            watermarks=watermarks,
-        )
-
-        assert watermarks == pytest.approx(
-            np.array(
-                [
-                    [0.6, 1, 1, 1],
-                    [0.1, 0.4, 0.3, 0.2],
-                    [0.1, 0.3, 0.2, 0.1],
-                    [0, 0, 0, 0],
-                    [0, 0, 0, 0],
-                    [0, 0, 0, 0],
-                ]
-            )
-        )
-
-
-class TestUpdateWatermarksNotEnough:
-    def test__first_captrue(self):
-
-        clocker = ac.Clocker()
-
-        watermarks = clocker.initial_watermarks_from_rows_and_total_traps(
-            rows=6, total_traps=1
-        )
-
-        electron_fractional_height = 0.5
-        enough = 0.7
-
-        watermarks = clock.updated_watermarks_from_capture_not_enough(
-            electron_fractional_height=electron_fractional_height,
-            watermarks=watermarks,
-            enough=enough,
-        )
-
-        assert watermarks == pytest.approx(
-            np.array([[0.5, 0.7], [0, 0], [0, 0], [0, 0], [0, 0], [0, 0]])
-        )
-
-    def test__new_highest_watermark(self):
-
-        watermarks = np.array(
-            [[0.5, 0.8], [0.2, 0.4], [0, 0], [0, 0], [0, 0], [0, 0]]
-        )
-
-        electron_fractional_height = 0.8
-        enough = 0.5
-
-        watermarks = clock.updated_watermarks_from_capture_not_enough(
-            electron_fractional_height=electron_fractional_height,
-            watermarks=watermarks,
-            enough=enough,
-        )
-
-        assert watermarks == pytest.approx(
-            np.array(
-                [[0.5, 0.9], [0.2, 0.7], [0.1, 0.5], [0, 0], [0, 0], [0, 0]]
-            )
-        )
-
-    def test__new_middle_watermark(self):
-
-        watermarks = np.array(
-            [[0.5, 0.8], [0.2, 0.4], [0.1, 0.3], [0, 0], [0, 0], [0, 0]]
-        )
-
-        electron_fractional_height = 0.6
-        enough = 0.5
-
-        watermarks = clock.updated_watermarks_from_capture_not_enough(
-            electron_fractional_height=electron_fractional_height,
-            watermarks=watermarks,
-            enough=enough,
-        )
-
-        assert watermarks == pytest.approx(
-            np.array(
-                [[0.5, 0.9], [0.1, 0.7], [0.1, 0.4], [0.1, 0.3], [0, 0], [0, 0]]
-            )
-        )
-
-    def test__new_lowest_watermark(self):
-
-        watermarks = np.array(
-            [[0.5, 0.8], [0.2, 0.4], [0.1, 0.3], [0, 0], [0, 0], [0, 0]]
-        )
-        electron_fractional_height = 0.3
-        enough = 0.5
-
-        watermarks = clock.updated_watermarks_from_capture_not_enough(
-            electron_fractional_height=electron_fractional_height,
-            watermarks=watermarks,
-            enough=enough,
-        )
-
-        assert watermarks == pytest.approx(
-            np.array(
-                [[0.3, 0.9], [0.2, 0.8], [0.2, 0.4], [0.1, 0.3], [0, 0], [0, 0]]
-            )
-        )
-
-    def test__multiple_traps(self):
-
-        watermarks = np.array(
-            [
-                [0.5, 0.8, 0.7, 0.6],
-                [0.2, 0.4, 0.3, 0.2],
-                [0.1, 0.3, 0.2, 0.1],
-                [0, 0, 0, 0],
-                [0, 0, 0, 0],
-                [0, 0, 0, 0],
-            ]
-        )
-        electron_fractional_height = 0.6
-        enough = 0.5
-
-        watermarks = clock.updated_watermarks_from_capture_not_enough(
-            electron_fractional_height=electron_fractional_height,
-            watermarks=watermarks,
-            enough=enough,
-        )
-        assert watermarks == pytest.approx(
-            np.array(
-                [
-                    [0.5, 0.9, 0.85, 0.8],
-                    [0.1, 0.7, 0.65, 0.6],
-                    [0.1, 0.4, 0.3, 0.2],
-                    [0.1, 0.3, 0.2, 0.1],
-                    [0, 0, 0, 0],
-                    [0, 0, 0, 0],
-                ]
-            )
-        )
-
-
-class TestElectronsCapturedInPixel:
-    def test__first_capture(self):
-
-        electrons_available = 2500  # --> electron_fractional_height = 0.5
-
-        clocker = ac.Clocker()
-        watermarks = clocker.initial_watermarks_from_rows_and_total_traps(
-            rows=6, total_traps=1
-        )
-
-        traps = [ac.Trap(density=10, lifetime=-1 / np.log(0.5))]
-
-        ccd_volume = ac.CCDVolume(
-            well_fill_beta=0.5, well_max_height=10000, well_notch_depth=1e-7
-        )
-
-        electrons_captured, watermarks = clock.electrons_captured_in_pixel(
-            electrons_available=electrons_available,
-            watermarks=watermarks,
-            traps=traps,
-            ccd_volume=ccd_volume,
-        )
-
-        assert electrons_captured == pytest.approx(5)
-        assert watermarks == pytest.approx(
-            np.array([[0.5, 1], [0, 0], [0, 0], [0, 0], [0, 0], [0, 0]])
-        )
-
-    def test__multiple_traps(self):
-
-        electrons_available = 3600  # --> electron_fractional_height = 0.6
-
-        watermarks = np.array(
-            [
-                [0.5, 0.8, 0.7],
-                [0.2, 0.4, 0.3],
-                [0.1, 0.3, 0.2],
-                [0, 0, 0],
-                [0, 0, 0],
-                [0, 0, 0],
-            ]
-        )
-
-        traps = [
-            ac.Trap(density=10, lifetime=-1 / np.log(0.5)),
-            ac.Trap(density=8, lifetime=-1 / np.log(0.2)),
-        ]
-
-        ccd_volume = ac.CCDVolume(
-            well_fill_beta=0.5, well_max_height=10000, well_notch_depth=1e-7
-        )
-
-        electrons_captured, watermarks = clock.electrons_captured_in_pixel(
-            electrons_available=electrons_available,
-            watermarks=watermarks,
-            traps=traps,
-            ccd_volume=ccd_volume,
-        )
-
-        assert electrons_captured == pytest.approx(
-            (0.1 + 0.06) * 10 + (0.15 + 0.07) * 8
-        )
-        assert watermarks == pytest.approx(
-            np.array(
-                [
-                    [0.6, 1, 1],
-                    [0.1, 0.4, 0.3],
-                    [0.1, 0.3, 0.2],
-                    [0, 0, 0],
-                    [0, 0, 0],
-                    [0, 0, 0],
-                ]
-            )
-        )
-
-    def test__not_enough__first_capture(self):
-
-        electrons_available = 2.5e-3  # --> electron_fractional_height = 4.9999e-4, enough=enough = 0.50001
-
-        clocker = ac.Clocker()
-        watermarks = clocker.initial_watermarks_from_rows_and_total_traps(
-            rows=6, total_traps=1
-        )
-
-        traps = [ac.Trap(density=10, lifetime=-1 / np.log(0.5))]
-
-        ccd_volume = ac.CCDVolume(
-            well_fill_beta=0.5, well_max_height=10000, well_notch_depth=1e-7
-        )
-
-        electrons_captured, watermarks = clock.electrons_captured_in_pixel(
-            electrons_available=electrons_available,
-            watermarks=watermarks,
-            traps=traps,
-            ccd_volume=ccd_volume,
-        )
-
-        assert electrons_captured == pytest.approx(0.0025)
-        assert watermarks == pytest.approx(
-            np.array(
-                [[4.9999e-4, 0.50001], [0, 0], [0, 0], [0, 0], [0, 0], [0, 0]]
-            )
-        )
-
-    def test__not_enough__multiple_traps(self):
-
-        electrons_available = 4.839e-4  # --> electron_fractional_height = 2.199545e-4, enough=enough = 0.5
-
-        watermarks = np.array(
-            [
-                [0.5, 0.8, 0.7],
-                [0.2, 0.4, 0.3],
-                [0.1, 0.3, 0.2],
-                [0, 0, 0],
-                [0, 0, 0],
-                [0, 0, 0],
-            ]
-        )
-
-        traps = [
-            ac.Trap(density=10, lifetime=-1 / np.log(0.5)),
-            ac.Trap(density=8, lifetime=-1 / np.log(0.2)),
-        ]
-
-        ccd_volume = ac.CCDVolume(
-            well_fill_beta=0.5, well_max_height=10000, well_notch_depth=1e-7
-        )
-
-        electrons_captured, watermarks = clock.electrons_captured_in_pixel(
-            electrons_available=electrons_available,
-            watermarks=watermarks,
-            traps=traps,
-            ccd_volume=ccd_volume,
-        )
-
-        assert electrons_captured == pytest.approx(
-            2.199545e-4 * (0.1 * 10 + 0.15 * 8)
-        )
-        assert watermarks == pytest.approx(
-            np.array(
-                [
-                    [2.199545e-4, 0.9, 0.85],
-                    [0.5 - 2.199545e-4, 0.8, 0.7],
-                    [0.2, 0.4, 0.3],
-                    [0.1, 0.3, 0.2],
-                    [0, 0, 0],
-                    [0, 0, 0],
-                ]
-            )
-        )
-
-
 class TestClocker:
     def test__express_matrix_from_rows(self):
         clocker = ac.Clocker(express=1)
@@ -620,29 +33,6 @@ class TestClocker:
         express_multiplier = clocker.express_matrix_from_rows(rows=12)
 
         assert express_multiplier == pytest.approx(np.triu(np.ones((12, 12))))
-
-    def test__initial_watermark_array__uses_rows_and_total_traps_to_set_size(
-        self,
-    ):
-        clock = ac.Clocker(sequence=None)
-
-        watermarks = clock.initial_watermarks_from_rows_and_total_traps(
-            rows=3, total_traps=1
-        )
-
-        assert (watermarks == np.zeros(shape=(3, 2))).all()
-
-        watermarks = clock.initial_watermarks_from_rows_and_total_traps(
-            rows=3, total_traps=5
-        )
-
-        assert (watermarks == np.zeros(shape=(3, 6))).all()
-
-        watermarks = clock.initial_watermarks_from_rows_and_total_traps(
-            rows=5, total_traps=2
-        )
-
-        assert (watermarks == np.zeros(shape=(5, 3))).all()
 
     def test__add_cti__single_pixel__compare_to_cpluspus_version(self):
         clocker = ac.Clocker(express=0)
